+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.2.0] - 2019-09-27
+
 ### Added
 
 - `http` and `grpc` `ServerOptions` to wrap a slice of options into a single element
@@ -21,4 +24,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial release
 
 
-[Unreleased]: https://github.com/sagikazarmark/kitx/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/sagikazarmark/kitx/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/sagikazarmark/kitx/compare/v0.1.0...v0.2.0